pyproject.toml
[project]
name = "webbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview[qt]>=6.0",
]
